refactor: alterando tipo de propriedade de long para decimal no contexto de preco

+++ src/CoffeeShop.Application/CoffeeShop.Application/UseCase/Pedido/GetTotalValorPedido/GetTotalValorPedidoUseCase.cs
@@ -4,7 +4,7 @@ namespace CoffeeShop.Application.UseCase.Pedido.GetTotalValorPedido
 {
     public class GetTotalValorPedidoUseCase : IGetTotalValorPedidoUseCase
     {
-        public long CalculateTotalValorPedido(List<PeiPedidoIten> listaPedido)
+        public decimal CalculateTotalValorPedido(List<PeiPedidoIten> listaPedido)
         {
             return listaPedido.Sum(item => item.PeiIntValorUnit * item.PeiIntQuantidade);
         }
